// src/types/overfast.ts
export interface components {
    schemas: {
// ...
        HeroKey: "ana" | "anran" | "ashe" | "baptiste" | "bastion" | "brigitte" | "cassidy" | "dva" | "domina" | "doomfist" | "echo" | "emre" | "freja" | "genji" | "hazard" | "hanzo" | "illari" | "jetpack-cat" | "junker-queen" | "junkrat" | "juno" | "kiriko" | "lifeweaver" | "lucio" | "mauga" | "mei" | "mercy" | "mizuki" | "moira" | "orisa" | "pharah" | "ramattra" | "reaper" | "reinhardt" | "roadhog" | "sigma" | "sierra" | "sojourn" | "soldier-76" | "sombra" | "symmetra" | "torbjorn" | "tracer" | "vendetta" | "venture" | "widowmaker" | "winston" | "wrecking-ball" | "wuyang" | "zarya" | "zenyatta";
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}